streaming_pipeline/requirements.txt
pandas>=1.3.0
requests>=2.26.0
boto3>=1.18.0
python-dotenv>=0.19.0
urllib3>=1.26.0 